# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MyCarQR API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile & subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [users]
      summary: Get dashboard summary stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ScanAlert:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        alertType:
          type: string
        message:
          type: string
        scannerLocation:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - vehicleId
        - alertType
        - isRead
        - createdAt
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [free, premium]
        vehicleCount:
          type: integer
        alertsThisMonth:
          type: integer
        isAdmin:
          type: boolean
        premiumExpiresAt:
          type: string
          nullable: true
      required:
        - userId
        - plan
        - vehicleCount
        - alertsThisMonth
        - isAdmin
    DashboardSummary:
      type: object
      properties:
        totalVehicles:
          type: integer
        activeQrCodes:
          type: integer
        totalAlertsThisMonth:
          type: integer
        unreadAlerts:
          type: integer
        documentsExpiringSoon:
          type: integer
        documentsExpired:
          type: integer
        averageSafetyScore:
          type: integer
        recentAlerts:
          type: array
          items:
            $ref: "#/components/schemas/ScanAlert"
      required:
        - totalVehicles
        - activeQrCodes
        - totalAlertsThisMonth
        - unreadAlerts
        - documentsExpiringSoon
        - documentsExpired
        - averageSafetyScore
        - recentAlerts
...