# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyCarQR API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfilePlan } from "./userProfilePlan";

export interface UserProfile {
  userId: string;
  email?: string;
  plan: UserProfilePlan;
  vehicleCount: number;
  alertsThisMonth: number;
  isAdmin: boolean;
  premiumExpiresAt?: string | null;
}